@@ -36,7 +36,7 @@ class ArtistController < ApplicationController
       end
 
       artist = Artist.find(params[:id])
-      artist.update_attributes(artist_params.merge(:updater_ip_addr => request.remote_ip, :updater_id => @current_user.id))
+      artist.update(artist_params.merge(:updater_ip_addr => request.remote_ip, :updater_id => @current_user.id))
 
       if artist.errors.empty?
         respond_to_success("Artist updated", :action => "show", :id => artist.id)
