@@ -1,4 +1,4 @@
-class Artist < ActiveRecord::Base
+class Artist < ApplicationRecord
   module UrlMethods
     module ClassMethods
       def find_all_by_url(url)
@@ -67,7 +67,7 @@ class Artist < ActiveRecord::Base
         elsif wiki_page.is_locked?
           errors.add(:notes, "are locked")
         else
-          wiki_page.update_attributes(:body => @notes, :ip_addr => updater_ip_addr, :user_id => updater_id)
+          wiki_page.update(:body => @notes, :ip_addr => updater_ip_addr, :user_id => updater_id)
         end
       end
     end
@@ -84,7 +84,7 @@ class Artist < ActiveRecord::Base
           self.class.where(:alias_id => id).update_all(:alias_id => nil)
           @alias_names.each do |name|
             a = Artist.find_or_create_by(:name => name)
-            a.update_attributes(:alias_id => id, :updater_id => updater_id)
+            a.update(:alias_id => id, :updater_id => updater_id)
           end
         end
       end
@@ -134,12 +134,12 @@ class Artist < ActiveRecord::Base
 
     def commit_members
       transaction do
-        self.class.where(:group_id => id).update_all(:group_id => nil)
+        self.class.unscoped.where(:group_id => id).update_all(:group_id => nil)
 
         if @member_names
           @member_names.each do |name|
             a = Artist.find_or_create_by(:name => name)
-            a.update_attributes(:group_id => id, :updater_id => updater_id)
+            a.update(:group_id => id, :updater_id => updater_id)
           end
         end
       end
