@@ -1,4 +1,4 @@
-class FlaggedPostDetail < ActiveRecord::Base
+class FlaggedPostDetail < ApplicationRecord
   belongs_to :post
   belongs_to :user
 
