@@ -1,4 +1,4 @@
-class PostVote < ActiveRecord::Base
+class PostVote < ApplicationRecord
   belongs_to :post
   belongs_to :user
 end
