@@ -8,7 +8,7 @@ class PostControllerTest < ActionController::TestCase
   end
 
   def update_post(post, params = {})
-    post.update_attributes({ :updater_user_id => 1, :updater_ip_addr => "127.0.0.1" }.merge(params))
+    post.update({ :updater_user_id => 1, :updater_ip_addr => "127.0.0.1" }.merge(params))
   end
 
   def create_default_posts
