@@ -8,7 +8,7 @@ class ArtistTest < ActiveSupport::TestCase
   end
 
   def update_artist(artist, params)
-    artist.update_attributes({ :updater_id => 1, :updater_ip_addr => "127.0.0.1" }.merge(params))
+    artist.update({ :updater_id => 1, :updater_ip_addr => "127.0.0.1" }.merge(params))
   end
 
   def test_normalize
