@@ -2,7 +2,17 @@ class HelpController < ApplicationController
   layout "default"
 
   def show
-    render "/help/#{params[:page].presence || "index"}"
+    page = params[:page].presence
+
+    if page.present?
+      sanitized_page = page.gsub(/[^a-z0-9_]/, '')
+
+      return head(:not_found) if sanitized_page != page
+    else
+      page = 'index'
+    end
+
+    render "/help/#{page}"
   rescue ActionView::MissingTemplate
     head :not_found
   end
