@@ -1,6 +1,6 @@
 window.DANBOORU_VERSION = {
-  major: 1,
-  minor: 13,
+  major: 6,
+  minor: 0,
   build: 0
 }
 
